azure_communication/compute_pip_requirements.txt
azureml-core==1.7.0
azureml-pipeline-core==1.7.0.post1
azureml-telemetry==1.7.0
azureml-defaults==1.7.0
azureml-interpret==1.7.0.post1
azureml-explain-model==1.7.0
azureml-automl-core==1.7.0.post1
azureml-automl-runtime==1.7.0.post2
azureml-train-automl-client==1.7.0.post1
azureml-train-automl-runtime==1.7.0.post1
azure-storage-blob
inference-schema
py-cpuinfo==5.0.0
pyodbc
pyarrow==0.17.0
